+++ src/duke/exception/DukeException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
